+++ package.json
@@ -38,6 +38,7 @@
     "@babel/preset-react": "^7.16.7",
     "autoprefixer": "^10.4.7",
     "babel-loader": "^8.2.5",
+    "body-parser": "^1.20.0",
     "css-loader": "^6.7.1",
     "html-webpack-plugin": "^5.5.0",
     "nodemon": "^2.0.16",
