+++ server/routers/mediaApi.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const path = require('path');
-const mediaController = require('./mediaController')
+const mediaController = require('../controllers/mediaController')
 
 const router = express.Router();
